move reaction JSON save function to cog
I've moved the portion of the code that saves a react role in the JSON file to a method in the ReactRole cog (__save_role_record). The variable naming in this commit probably isn't too great, but I don't understand what it does yet well enough to improve that. I did not hardcode the JSON file path, so that it can easily be configured.