fix fun.py
All errors, obvious bugs and mypy annotations in fun.py have been fixed. The dadjokes have been moved into a json file to keep things clean, and are loaded each time the command is run. This will cause IO wait, and could possibly be moved to an asyncio executor, but it will allow the dad jokes to be modified on the fly while the bot is running.